devshouse-25/middlewares/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
python-dotenv>=1.0.0
authlib>=1.2.0
python-jose[cryptography]>=3.3.0
cryptography>=42.0.0
httpx>=0.27.0
pydantic-settings>=2.2.0
